Extract `PreservationTable` from `StripeSM`
This extracts `PreservationTable` as a superclass. The new class represents
the responsibility of tracking documents that need to be preserved to avoid
being overwritten, such as pinned documents. Docstrings are added to better
document how this process works. The `StripeSM::evac_range` method is left
with that class because it is not part of the evacuation data structure
despite the name. It should be refactored and renamed in a future PR.
